feat: update Job class and add JobType interface
Refactored the Job class in the core jobs module to remove the WorkingSteps related methods and replace them with JobType related methods. JobType is a new interface added that represents different types of jobs. This update aims to improve flexibility and scalability in the management of various job types in the game. The Job class now utilizes Gson for more straightforward object serialization and deserialization.