feat: update job data saving and Gson setup
Modified the way job data is saved to rely on a Job object instead of a String object. The Gson object now includes type adapters for Location and ItemStack, to better handle these types of data. Construction and non-existence checks for both player and job data directories have also been added, increasing the robustness of the system.