refactor: job handling and update menu interaction
The job handling code in AllJobsMenu has been refactored to reduce redundancy and improve readability. Now, it uses optional chaining to directly open the job admin menu if a job is found. Additionally, an if-statement has been added to the EasyPaginatedMenu to prevent interactions with BORDER_GLASS items. This stops potential errors and unexpected behavior from occurring in the menu.