feat: add Vault integration to the plugin
The Vault API has been added as a dependency in pom.xml. In the plugin.java file, the Vault economy system has been set up and integrated. If Vault is not installed, the plugin will be disabled. The plugin.yml file has been updated accordingly to include Vault as a soft dependency.