feat: remove onPlayerChat event in CreateJobCommand
The changes have eliminated the onPlayerChat event from the CreateJobCommand .java class. This is part of an effort to streamline the command creation process. This event handler was handling player chat during the job creation, but its functionality has been removed in favour of a more direct command interaction approach.