refactor: hologram creation and assignment in commands
The hologram creation code was duplicated and over-complicated in DoCommand and MeCommand. Now it's more streamlined and encapsulated in a separate method to improve readability and maintainability. The check to create a new hologram is also performed inside the method for better responsibility separation. Adjustments were also made in FollowingHologram to handle the assignment to player holograms whenever a new hologram is attached to an entity.