feat: implement job addition and removal in PlayerData
The commit adds methods to add and remove a job from a player's list of jobs in the PlayerData class. It also updates the Job class to reflect these changes - now when a player is added or removed from a job, this change is reflected in their PlayerData. Furthermore, a getJob method has been implemented in the Job class.