feat: add PlaceholderAPI support and respective dependencies
A new class `RPUExpansion` is added to handle PlaceholderAPI integration. The `pom.xml` file has been updated to include PlaceholderAPI in the dependencies, with a new repository added for this. In `plugin.yml`, PlaceholderAPI is added as a soft dependency and additional permissions related to PlaceholderAPI usage are also defined.