feat: add job removal functionality and refactor loading process
Code now includes the ability for jobs in the RPUniverse core to be removed, which is facilitated by the addition of the 'remove' method in Job.java. The process of loading player data upon game start has also been optimized by restructuring the order of system initialization in RPUniverse.java and improving player data retrieval in PlayerData.java and JSONDataHandler.java.