feat: position setters now reopen menus
Modified the Position setters in the Position class to reopen menus upon value change. This improve how the menus respond and update in realtime when job positions are altered, enhancing the in-app user interaction and experience. appropriated modifications have been performed on Job and AllJobsMenu files to accommodate these changes.