feat: Added JobsHandler event registration and PlayerJob modifications
Added code in RPUniverse.java to register the JobsHandler events to the server plugin manager. In PlayerData.java, introduced new methods to set and get 'selectedPlayerJob' and also implemented getPlayerJobs method. Implemented an updateBossBar method and a onPlayerJoin event in JobsHandler.java to manage player's job updates and job assignment on joining respectively.