feat: update BossBar and PlayerData with job status & added message when player doesn't have a job
Added utilities to PlayerData class to update the BossBar when a player's job status changes. This allows for job selection, removal, switching, and mode change to be reflected in the BossBar. Additionally, the code for updating the BossBar has been moved from the JobsHandler to the PlayerData class for better organization and simplicity.