feat: add player kick and position assignment functionalities
Two new methods were added to the Job class. The first one, `canPlayerKickPlayer`, allows a player to kick another player based on their positions. The second one, `getAllPositionsPlayerCanAssign`, provides a list of all positions that a player can assign to, which could be empty if there are no available positions.